MOD3Project
Create a sample database and table using phpMyAdmin in your Microsoft Azure account.
Populate the database with data that is related to your website. This must include at least 4 columns, including a primary key.
Construct a simple SELECT query with 3 or more columns.
Construct a web page with responsive styling that displays the results of a query.